+++ dashboard/src/nav/Navbar.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import {FaUser, FaCog, FaSync} from 'react-icons/fa'; // Import User, Cog and Restart icons
+import {FaUser, FaCog, FaSync, FaSpinner, FaCheck, FaTimes} from 'react-icons/fa'; // Import User, Cog and Restart icons
 import { NavLink } from 'react-router-dom';
 import './Navbar.css'
 import orcLogo from '/orc_favicon.svg'
@@ -7,12 +7,9 @@ import MessageBox from "../messageBox.jsx";
 import api from "../api/api.js";
 import { useAuth } from "../auth/useAuth.jsx";
 
-
-
-const Navbar = ({requiresRestart, setRequiresRestart, setIsLoading}) => {
+const Navbar = ({requiresRestart, setRequiresRestart, setIsLoading, videoRunState}) => {
     const [isOpen, setIsOpen] = useState(false); // track if navbar is open / closed
     const [settingsOpen, setSettingsOpen] = useState(false); // track if settings menu is open
-
     const { logout } = useAuth();
     const handleToggle = (openState, setOpenState) => {
       setOpenState(!openState); // Toggles the `isOpen` state
@@ -24,15 +21,11 @@ const Navbar = ({requiresRestart, setRequiresRestart, setIsLoading}) => {
 
     const handleUserButtonClick = async () => {
         await logout();
-        // api.post("/security/logout")
-
-      // alert('User login functionality to be implemented.');
     };
 
     const handleSettingsClick = () => {
         setSettingsOpen(!settingsOpen);
     };
-
 
     const handleRestartClick = () => {
       setIsLoading(true);
@@ -41,13 +34,62 @@ const Navbar = ({requiresRestart, setRequiresRestart, setIsLoading}) => {
       api.post("/updates/shutdown")
     }
 
-    return (
+  const getStatusIcon = (status) => {
+    switch (status) {
+      case 1:
+        return
+      case 3:
+        return <span><FaSpinner style={{color: "white", animation: "spin 1s linear infinite"}}/> </span>// Spinner for processing
+      case 4:
+        return <span><FaCheck style={{
+          color: "green",
+          filter: "drop-shadow(0px 0px 1px white)",
+        }}/> </span>; // Success
+      case 5:
+        return <span><FaTimes style={{
+          color: "red",
+          filter: "drop-shadow(0px 0px 1px white)",
+        }}/> error</span>; // Error
+      default:
+        return
+    }
+  };
+  const getSyncStatusIcon = (status) => {
+    switch (status) {
+      case 1:
+        return
+      case 5:
+        return <span><FaSpinner style={{color: "white", animation: "spin 1s linear infinite"}}/> </span>// Spinner for syncing
+      case 2:
+        return <span><FaCheck style={{
+          color: "cadetblue",
+          filter: "drop-shadow(0px 0px 1px white)",
+        }}/> </span>; // Error
+
+      case 4:
+        return <span><FaTimes style={{
+          color: "red",
+          filter: "drop-shadow(0px 0px 1px white)",
+        }}/> </span>; // Error
+      default:
+        return
+    }
+  };
+
+  return (
         <>
             <nav className='navbar navbar-dark'>
                 <div className='container-fluid'>
                     <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation" onClick={() => handleToggle(isOpen, setIsOpen)}>
                         <span className="navbar-toggler-icon"></span>
                     </button>
+                  <div className="navbar-message" style={{ marginRight: 'auto', marginLeft: '10px' }}>
+                    {videoRunState?.video_file !== "" && (
+                      <span style={{ fontWeight: 'bold', position: 'absolute', overflow: 'hidden', zIndex: 0, width: '700px', whiteSpace: 'nowrap', display: 'inline-block', textOverflow: 'ellipsis'}}>
+                        {getStatusIcon(videoRunState.status)} {getSyncStatusIcon(videoRunState.sync_status)} {videoRunState.video_file} - {videoRunState.message}
+                      </span>
+                    )}
+                  </div>
                     <div className="navbar-right">
                         <MessageBox/>
                         <FaSync
@@ -71,16 +113,22 @@ const Navbar = ({requiresRestart, setRequiresRestart, setIsLoading}) => {
                 <div className="sidebar-header">
                     <a className='navbar-brand' href="#">
                         <img src={orcLogo} alt="ORC Logo" width="30" height="30" className="d-inline-block align-text-top"/>
-                    {' '} NodeORC
+                    {' '} ORC-OS
                     </a>
                     <button className="close-button" onClick={() => {handleToggle(isOpen, setIsOpen);}}>
                         &times;
                     </button>
                 </div>
                 <ul className="sidebar-nav">
-                    <hr/>
                     <li className="sidebar-brand" style={{fontSize: "25px"}}>Menu</li>
                     <hr/>
+                    <li>
+                      <NavLink
+                        className={({ isActive }) => isActive ? "sidebar-link active" : "sidebar-link"}
+                        to="/" onClick={handleClose}>
+                        Home
+                      </NavLink>
+                    </li>
 
                     <li>
                         <NavLink
@@ -90,17 +138,17 @@ const Navbar = ({requiresRestart, setRequiresRestart, setIsLoading}) => {
                         </NavLink>
                     </li>
                     <li>
+                      <NavLink
+                        className={({ isActive }) => isActive ? "sidebar-link active" : "sidebar-link"}
+                        to="/log" onClick={handleClose}>
+                        Log file
+                      </NavLink>
+                    </li>
+                    <li>
                         <NavLink
                           className={({ isActive }) => isActive ? "sidebar-link active" : "sidebar-link"}
                           to="/camera_aim" onClick={handleClose}>
                             Aim your camera in the field
-                        </NavLink>
-                    </li>
-                    <li>
-                        <NavLink
-                            className={({ isActive }) => isActive ? "sidebar-link active" : "sidebar-link"}
-                            to="/" onClick={handleClose}>
-                        Home
                         </NavLink>
                     </li>
                     <li>
@@ -124,10 +172,6 @@ const Navbar = ({requiresRestart, setRequiresRestart, setIsLoading}) => {
                             Videos
                         </NavLink>
                     </li>
-                    <hr/>
-                    <li className="sidebar-brand" style={{fontSize: "25px"}}>Settings</li>
-                    <hr/>
-
                 </ul>
             </div>
             {isOpen && <div className="sidebar-overlay" onClick={() => handleToggle(isOpen, setIsOpen)}></div>}