+++ orc_api/log.py
@@ -1,15 +1,16 @@
 """Logging utilities."""
 
+import asyncio
 import logging
 import logging.handlers
 import os
 import sys
-from datetime import datetime
+
+import anyio
+from fastapi import WebSocket
 
 from orc_api import LOG_DIRECTORY, __home__, __version__
 
-timestr = datetime.now().strftime("%Y%m%dT%H%M%S")
-datestr = datetime.now().strftime("%Y%m%d")
 FMT = "%(asctime)s - %(name)s - %(module)s - %(levelname)s - %(message)s"
 
 
@@ -51,28 +52,40 @@ def setuplog(
     console.setFormatter(logging.Formatter(fmt))
     logger.addHandler(console)
     if path is not None:
-        if append is False and os.path.isfile(path):
-            os.unlink(path)
-        add_filehandler(logger, path, log_level=log_level, fmt=fmt)
+        # if append is False and os.path.isfile(path):
+        #     os.unlink(path)
+        add_filehandler(logger, path, log_level=log_level, fmt=fmt, backupCount=10)
     logger.info(f"ORC-OS version: {__version__}")
 
     return logger
 
 
-def add_filehandler(logger, path, log_level=20, fmt=FMT):
+def add_filehandler(logger, path, log_level=20, fmt=FMT, backupCount=0):
     """Add file handler to logger."""
     if not os.path.isdir(os.path.dirname(path)):
-        print(path)
         os.makedirs(os.path.dirname(path))
-    isfile = os.path.isfile(path)
-    ch = logging.FileHandler(path)
+    if backupCount > 0:
+        ch = logging.handlers.RotatingFileHandler(path, backupCount=backupCount, encoding="utf-8")
+        # Force a rollover to create a new log file on application startup
+        ch.doRollover()
+
+    else:
+        if os.path.isfile(path):
+            os.unlink(path)
+        ch = logging.FileHandler(path, encoding="utf-8")
     ch.setFormatter(logging.Formatter(fmt))
     ch.setLevel(log_level)
     logger.addHandler(ch)
-    if isfile:
-        logger.debug(f"Appending log messages to file {path}.")
-    else:
-        logger.debug(f"Writing log messages to new file {path}.")
+
+
+def remove_file_handler(logger, name_contains="hello_world"):
+    """Remove file handler with name_contains in file name from logger."""
+    for handler in logger.handlers:
+        if isinstance(handler, logging.FileHandler):
+            if name_contains in handler.baseFilename:
+                logger.removeHandler(handler)
+                handler.close()
+                logger.debug(f"Removed file handler to {handler.baseFilename} from logger.")
 
 
 def start_logger(verbose, quiet, log_path=None):
@@ -87,10 +100,51 @@ def start_logger(verbose, quiet, log_path=None):
         quiet = 1
     else:
         quiet = 0
-    logfile = os.path.abspath(os.path.join(log_path, datestr, f"orc-os-{timestr}.log"))
+    logfile = os.path.abspath(os.path.join(log_path, "orc-os.log"))
     log_level = max(10, 30 - 10 * (verbose - quiet))
     logger = setuplog(name="ORC-OS", path=logfile, log_level=log_level)
     return logger
+
+
+def get_last_lines(fn: str, count: int = 10):
+    """Get the last `count` lines from a file."""
+    if not os.path.exists(fn):
+        raise FileNotFoundError(f"Log file not found: {fn}")
+
+    # ensure count is limited
+    restricted_count = min(count, 500)
+    with open(fn, "rb") as f:
+        f.seek(0, os.SEEK_END)  # Move to the end of the file
+        buffer = bytearray()
+        lines_found = 0
+        # Start reading backward
+        while f.tell() > 0 and lines_found <= restricted_count:
+            f.seek(-1, os.SEEK_CUR)  # Step back one character
+            byte = f.read(1)  # Read one character
+            if byte == b"\n":  # Check for new line
+                lines_found += 1
+                if lines_found > restricted_count:
+                    break
+            buffer.extend(byte)
+            f.seek(-1, os.SEEK_CUR)  # Step back again to continue
+    return buffer[::-1].decode("utf-8")  # Reverse the buffer to get the correct order
+
+
+async def stream_new_lines(websocket: WebSocket, fn: str):
+    """Stream new lines from a file as they are written."""
+    if not os.path.exists(fn):
+        raise FileNotFoundError(f"File {fn} does not exist.")
+
+    async with await anyio.open_file(fn, "r") as f:
+        # with open(fn, "r") as f:
+        # Move to the end of the file
+        f.seek(0, 2)
+        while True:
+            line = await f.readline()
+            if line:
+                await websocket.send_text(line)  # Send the new line to the WebSocket
+            else:
+                await asyncio.sleep(0.1)  # Wait before trying to read more lines
 
 
 if "ALEMBIC_RUNNING" not in os.environ:
@@ -98,4 +152,4 @@ if "ALEMBIC_RUNNING" not in os.environ:
 else:
     logger = None
 
-__all__ = ["logger"]
+__all__ = ["logger", "get_last_lines", "stream_new_lines"]