+++ tests/test_schemas/test_video_schemas.py
@@ -18,7 +18,7 @@ def test_video_run(
 ):
     monkeypatch.setattr("orc_api.schemas.video.get_session", lambda: session_video_config)
     assert video_response.status == models.VideoStatus.NEW
-    video_response.run(base_path=sample_data.get_hommerich_pyorc_files())
+    video_response.run(session=session_video_config, base_path=sample_data.get_hommerich_pyorc_files())
     assert video_response.status == models.VideoStatus.DONE
     assert len(video_response.get_netcdf_files(base_path=sample_data.get_hommerich_pyorc_files())) > 0
     assert video_response.get_discharge_file(base_path=sample_data.get_hommerich_pyorc_files()) is not None
@@ -46,6 +46,9 @@ def test_video_run_daemon_shutdown(tmpdir, video_response_no_ts, session_video_c
     """Mock running of video in order to test if shutdown is handled correctly."""
     monkeypatch.setattr("orc_api.schemas.video.get_session", lambda: session_video_config)
 
+    class ShutdownException(Exception):
+        pass
+
     def mock_velocity_flow(**kwargs):
         return None
 
@@ -57,6 +60,8 @@ def test_video_run_daemon_shutdown(tmpdir, video_response_no_ts, session_video_c
 
     def mock_subprocess_call(*args, **kwargs):
         print("Shutdown called!")
+        raise ShutdownException("Simulating a shutdown")
+
         return None
 
     mock_shutdown = mock.Mock(side_effect=mock_subprocess_call)
@@ -64,8 +69,8 @@ def test_video_run_daemon_shutdown(tmpdir, video_response_no_ts, session_video_c
     monkeypatch.setattr("orc_api.schemas.video.velocity_flow", mock_velocity_flow)
     monkeypatch.setattr("orc_api.schemas.video.VideoResponse.update_timeseries", mock_update_timeseries)
     monkeypatch.setattr("orc_api.schemas.video.VideoResponse.sync_remote", mock_update_timeseries)
-
-    video_response_no_ts.run(session=session_video_config, base_path=tmpdir, shutdown_after_task=True)
+    with pytest.raises(ShutdownException):
+        video_response_no_ts.run(session=session_video_config, base_path=tmpdir, shutdown_after_task=True)
     # test if mock shutdown is called once
     assert mock_shutdown.call_count == 1
 
@@ -152,7 +157,7 @@ def test_video_sync_real_server(session_video_with_config, video_response, monke
         url=os.getenv("LIVEORC_URL"),
         user=os.getenv("LIVEORC_EMAIL"),
         password=os.getenv("LIVEORC_PASSWORD"),
-        retry_timeout=10,
+        retry_timeout=60,
     )
     tokens = callback_create.get_tokens().json()
     new_callback_dict = callback_create.model_dump(exclude_none=True, mode="json", exclude={"id", "password", "user"})
@@ -162,7 +167,7 @@ def test_video_sync_real_server(session_video_with_config, video_response, monke
             "token_access": tokens["access"],
             "token_refresh": tokens["refresh"],
             "token_expiration": callback_create.get_token_expiration(),
-            "retry_timeout": 10.0,
+            "retry_timeout": 60.0,
         }
     )
     new_callback_url = models.CallbackUrl(**new_callback_dict)